Fix incorrect handling of dinucleotide repeats, add test
Was previously not keeping track of the running length diff.  Also the
pattern match was broken.  Basically it was only working in scenarios
where a single dinucleotide repeat block needed correction in a
window.  This is the most common case, but I am sure we were still
seeing errors because of this.

[Tracking bug 22690]

[git-p4: depot-paths = "//depot/software/bioinformatics/tools/GenomicConsensus/": change = 122988]