Fix bug 23401: Lawrence's quiver crash / memory leak in HDF5 use
The fix is the addition of a flag --disableHdf5ChunkCache,
which disables the HDF5 chunk cache on worker processes.  The
default chunk cache settings result in a 1MB chunk cache per
dataset.  For Lawrence's job with ~6000 reference groups, this
entailed ~30GB of memory used per worker process!

The longer term fix is likely to involve reworking the cmp.h5
file format, but in the interim, I might consider
automatically disabling the chunk cache for files with >250
reference groups, say.  Another possibility would be using an
LRU cache in the CmpH5Reader to maintain the datasets in use
at any time, actively closing old datasets.

[git-p4: depot-paths = "//depot/software/bioinformatics/tools/GenomicConsensus/": change = 124383]