Partial fix for frequent convergence failures at low coverage
If we are about to apply a set of mutations that would
introduce a cycle (revisiting a template we've already seen),
apply a subset of the mutations instead.  The code as I've
written it could still actually run into a cycle but it is
unlikely.

This fixes about 50% of the convergence failures I have been
observing on the 10x human dataset.

[Tracking bug 23893]

[git-p4: depot-paths = "//depot/software/bioinformatics/tools/GenomicConsensus/": change = 129083]