Fix bug 29116 (quiver "IOError: bad message length" followed by freeze)
There was no coverage against chr1, so it was getting
no-called, which meant a struct was getting sent over the
multiprocessing pickle queue that contained an array of zeros
the length of chr1.  And it was being encoded as a numpy array
of uint64, as opposed to uint8, so it was needlessly consuming
way too much memory, somehow triggering failure in
multiprocessing.

Fix is to encode as uint8 array.

[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/tools/GenomicConsensus/": change = 165287]