Remove abstraction-breaking Reader accession
The previous implementation went through each resource reader in the current
AlignmentSet, assembling tStart tEnd pairs for relevant subreads. Now that a
pbi file is a requirement for using GenomicConsensus and due to a few DataSet
API improvements, we no longer have to break this abstraction.

Instead, we do a few pbindex table queries and we're done.

Updated requirements to pbcore >= 1.2.8