arrow: add --reportEffectiveCoverage flag
Use of the flag --reportEffectiveCoverage will add the
"effectiveCoverage" attribute to each variant reported in the GFF
output.  This metric represents the reference-site-coverage actually
used to make the consensus call---after filters, after failures to add
reads (bad ZScores or Alpha/Beta mismatch errors in the consensus
machinery, for example).

This should not affect the production mode in any way.